move StitchGroup into lib.stitch_plan
