+++ lib/stitch_plan/__init__.py
@@ -3,6 +3,8 @@
 # Copyright (c) 2010 Authors
 # Licensed under the GNU GPL version 3.0 or later.  See the file LICENSE for details.
 
-from .stitch_plan import patches_to_stitch_plan, StitchPlan, ColorBlock
+from .stitch_plan import patches_to_stitch_plan, StitchPlan
+from .color_block import ColorBlock
+from .stitch_group import StitchGroup
 from .stitch import Stitch
 from .read_file import stitch_plan_from_file
